+++ integration-test/end-to-end.js
@@ -1,4 +1,4 @@
-const postMessage = require("../src/post-message");
+const { postMessage } = require("../src/slack-api");
 const buildMessage = require("../src/build-message");
 
 (async () => {
@@ -13,7 +13,7 @@ const buildMessage = require("../src/build-message");
 
   const message = buildMessage(process.env.CHANNEL, process.env.TEXT, {
     as_user: false,
-    icon_emoji: ":fire:"
+    icon_emoji: ":fire:",
   });
   const result = await postMessage(
     process.env.BOT_USER_OAUTH_ACCESS_TOKEN,
